Restructure specs as per standard rspec-rails
Specifically we:

 - Move Clean Architecture into spec/lib
 - Move helpers into spec/support
 - Move DB or external dep dependent rspec config into rails_helper
