+++ spec/lib/hackney/income/stored_tenancies_gateway_spec.rb
