+++ spec/lib/hackney/universal_housing/client_spec.rb
