+++ spec/spec_helper.rb
@@ -1,6 +1,5 @@
-require 'webmock/rspec'
-require 'tenancy_helper'
-require 'universal_housing_helper'
+require 'support/faker'
+require 'support/tenancy_helper'
 
 # This file was generated by the `rails generate rspec:install` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
@@ -19,7 +18,6 @@ require 'universal_housing_helper'
 # See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 RSpec.configure do |config|
   config.include TenancyHelper
-  config.include UniversalHousingHelper
 
   # rspec-expectations config goes here. You can use an alternate
   # assertion/expectation library such as wrong or the stdlib/minitest
