+++ spec/support/constants.rb
@@ -1,0 +1,1 @@
+require Rails.root.join('lib/hackney/income')
