+++ spec/support/faker.rb
@@ -1,0 +1,1 @@
+require 'faker'
