+++ spec/support/webmock.rb
@@ -1,0 +1,1 @@
+require 'webmock/rspec'
